chore: changed class to static
